+++ testdata/intentions/statements/convertStatementToIfIntention/ComplexExpression.test
@@ -1,0 +1,11 @@
+package main
+
+func main() {
+    1+2==3||tr<caret>ue
+}
+-----
+package main
+
+func main() {
+    if 1+2==3||true{<caret>}
+}
